Crate nib

Modules

accel
A module defining interfaces for an acceleration structure

camera
A module defining the generic interface for cameras and providing interfaces for various camera types

cli
This module handles everything related to the CLI interface, such as arguments, "UI", and control flow.

hittable
A module defining an interface for objects in a scene that a ray can intersect with

image_exporter
Export a framebuffer to some image format

integrator
The integrator module defines several integrators, and also provides an interface for generic integrators that can be extended so that other integrators can be easily added.

material
Defines the interface for materials that are associated with geometric objects. This interface gives us texture and lighting information by defining a BSDF function for a particular geometric object.

math
Utility math functions

ray
Defines a Ray type with convenience methods

renderer
The renderer implementation, which is pretty much the entire app

sampler
The sampler interface

scene
A module providing definitions and methods for a scene. A scene contains most of the relevant information for an integrator, and it provides information such as lights, geometric objects, camera information, amongst other things. The scene primarily interacts with the deserializer and the integrator.

types
Defines generic numeric types for the integrator so that operations can be done with generic integers or floating point numbers.

Statics

ALLOC

Functions

main