Module nib::accel

A module defining interfaces for an acceleration structure

This module provides the generic interface for acceleration structures as well as implementations of various acceleration structures.

Re-exports

pub use list::ObjectList;

Modules

list
The "list" acceleration structure for computing intersections.

Structs

AccelRecord
The result of a query to an acceleration structure.

TexturedRef
A reference to an object that consists of a geometric form and a texture

Enums

AccelError
An error associated with acceleration structures

Traits

Accel
The Accel trait is a generic trait for acceleration structures in the renderer.

Type Definitions

AccelResult
A result that can return an AccelError