Trait nib::accel::Accel

pub trait Accel<T: GenFloat>: Debug + Send + Sync {
    fn collision(&self, ray: &Ray<T>) -> Option<AccelRecord<T>>;
}

The Accel trait is a generic trait for acceleration structures in the renderer.

It provides a simple interface which allows the caller to determine if an incoming ray collided with some object in the scene.

Required methods

fn collision(&self, ray: &Ray<T>) -> Option<AccelRecord<T>>

Return whether the incoming ray collided with any of the objects in the scene

Loading content...