Trait nib::camera::Camera

pub trait Camera: Debug + Send + Sync {
    // Required method
    fn to_ray(&self, u: f32, v: f32) -> Ray;
}

The generic interface for a camera type

A camera simply needs to convert u, v coordinates to a 3D ray.

Required Methods

fn to_ray(&self, u: f32, v: f32) -> Ray

Convert (u, v) pixel coordinates to a ray in 3D space

This method expects (u, v) coordinates that lie in the unit plane [0, 1]. It also expects an aspect ratio, which is just nx / ny, where nx and ny are the horizontal and vertical pixels, respectively.