Function nib::cli::dispatch_scene_parse

pub fn dispatch_scene_parse(
    path: &PathBuf, 
    ext: Option<&str>
) -> Result<Scene<f32>>

Parse the input scene file based on the file extension

Because I am allowing multiple types of files for the scene files (e.g. JSON or YAML), and also may change the formats in the future, we have a wrapper method to parse the scene file into a scene struct. Right now it attempts to determine the file type from the path's extension.

Right now we support the following filetypes for the scene description:

yaml

json

json5 (which is a superset of json)

RON (Rusty Object Notation)

I would recommend using RON since it's the most expressive give that we are using Rust data structures, and it has full support for all of serde's data types, which is what we're using to serialize.