Module nib::hittable

A module defining an interface for objects in a scene that a ray can intersect with

This module defines interfaces for all of the geometry in the renderer (which are all things that a ray can intersect with), as well as interfaces for "infinite" objects that lie outside the scene (like skyboxes and ambient lighting), as well as interfaces for acceleration structures so that they can be used generically. It doesn't really matter, as long as you can yield which object was hit.

Re-exports

pub use sphere::Sphere;

Modules

sphere
An implementation of the sphere primitive

Structs

HitRecord
Information pertaining to a ray intersection

Textured
The struct for some object in the scene that can be intersected geometrically that also provides a BSDF function for texture.

Enums

SerializedHittable
The different types of Hittable types that can be used as input objects

Constants

_IMPL_DESERIALIZE_FOR_SerializedHittable

_IMPL_SERIALIZE_FOR_SerializedHittable

Traits

Hittable
An interface for any object that can intersect with a ray coming from the camera