Trait nib::hittable::AccelHittable

pub trait AccelHittable<T: GenFloat>: Debug + Send + Sync {
    fn hit(&self, ray: &Ray<T>, inverse_dir: Vector3<T>) -> bool;
}

An interface for an object that is used as part of an acceleration structure

This method is automatically implemented for any struct that implements Hittable. This differs from the Hittable trait in that these objects simply need to report whether they were hit or not, not the particular point where the ray intersects the object.

Required methods

fn hit(&self, ray: &Ray<T>, inverse_dir: Vector3<T>) -> bool

Whether the objcet was hit

This method simply returns whether the object intersected with the ray. It requires the incoming ray and the inverse direction of the ray as parameters. The inverse ray is precomputed for efficiency, since it's a calculation that will be repeated for every potential ray intersection.

Loading content...