Module nib::hittable::triangle

An implementation of the triangle primitive hittable

The triangle is probably one of the most widely used geometric primitives in three dimensional rendering and modeling, as most OBJ files are defined in terms of triangles.

Structs

Triangle
A geometric triangle

TriangleParameters
A geometric triangle

Enums

TriangleHandedness
The "handedness" of the coordinate system used to define the triangle

Constants

_IMPL_DESERIALIZE_FOR_Triangle

_IMPL_DESERIALIZE_FOR_TriangleHandedness

_IMPL_DESERIALIZE_FOR_TriangleParameters

_IMPL_SERIALIZE_FOR_Triangle

_IMPL_SERIALIZE_FOR_TriangleHandedness

_IMPL_SERIALIZE_FOR_TriangleParameters

Functions

default_handedness
A helper method for serde to infer the default handedness of a triangle