Module nib::image_exporter

Export a framebuffer to some image format

Once you have rendered an image, you have a buffer of RGB values. This module provides interfaces to export that framebuffer to a file, such as a PNG or PPM.

Structs

PNGExporter

PPMExporter
Export a framebuffer to the PPM image format

Enums

ExporterError
The possible errors that can arise when exporting a framebuffer

OuputType
An enum type describing the possible output filetypes for the resulting image

Traits

FramebufferExporter
Something that can export a framebuffer of PixelValues to some other format

FramebufferExporterBase
The "base" trait for a FrameBufferExporter

Type Definitions

ExporterResult
A result that can return an ExporterError