Trait nib::image_exporter::FramebufferExporter

pub trait FramebufferExporter {
    fn export<T: GenFloat>(
        &self, 
        buffer: &[Vector3<T>], 
        path: &Path
    ) -> Result<(), ExporterError>;
}

Something that can export a framebuffer of PixelValues to some other format

Users should use this trait, implementors should use the FramebufferExporterBase trait, which abstracts away some common methods, such as converting a floating point value to an integer RGB value.

Required methods

fn export<T: GenFloat>(
&self,
buffer: &[Vector3<T>],
path: &Path
) -> Result<(), ExporterError>

Export a buffer of floating point pixel values to some other format

This method expects a framebuffer of pixel values between 0 and 1, that haven't been converted to some specific color format yet.

Loading content...