Module nib::integrator

The integrator module defines several integrators, and also provides an interface for generic integrators that can be extended so that other integrators can be easily added.

Re-exports

pub use normal::Normal;

pub use whitted::Whitted;

Modules

normal
The normal integrator is a diagnostic integrator that simply returns the normal values of the surfaces it hits.

whitted
An implementation of the Whitted/direct lighting rendering technique. This is the "classic" ray-tracing technique. It was developed by Turner Whitted in 1980, in a paper titled "An Improved Illumination Model for Shaded Display."

Structs

RenderParams
A struct representing the various options and parameters that can be configured for the render method in a particular integrator

Enums

SerializedIntegrator

Constants

_IMPL_DESERIALIZE_FOR_SerializedIntegrator

_IMPL_SERIALIZE_FOR_SerializedIntegrator

Traits

Integrator
A trait that defines an integrator. An integrator defines the operations that are responsible for taking input data for a given pixel, and calculating the output colors at each pixel.