Module nib::integrator::normal

The normal integrator is a diagnostic integrator that simply returns the normal values of the surfaces it hits.

This is useful to ensure that normals and surface intersections are being calculated correctly, and that the geometry of the scene is correct.

Structs

Normal
The parameters for the Normal integrator

Constants

_IMPL_DESERIALIZE_FOR_Normal

_IMPL_SERIALIZE_FOR_Normal