nib::integrator
Trait Integrator

pub trait Integrator:
    Debug
    + Send
    + Sync {
    // Required method
    fn render(&self, params: RenderParams<'_, '_, '_>) -> Vector3<f32>;
}

A trait that defines an integrator. An integrator defines the operations that are responsible for taking input data for a given pixel, and calculating the output colors at each pixel.

Required Methods

fn render(&self, params: RenderParams<'_, '_, '_>) -> Vector3<f32>

Calculate the color value for a particular pixel, given a reference to the scene.

Given certain input parameters this method calculates the color values at a particular point.