Trait nib::material::BSDF

pub trait BSDF: Debug + Send + Sync {
    fn scatter(
        &self, 
        s: &mut dyn Sampler<f32>, 
        ray: &Ray, 
        hit_record: &HitRecord
    ) -> BSDFRecord;
}

This trait defines some sort of object that can specify how light is scattered when the material is hit.

This interface provides one method: the scatter function, which will return a BSDFRecord

Required methods

fn scatter(
&self,
s: &mut dyn Sampler<f32>,
ray: &Ray,
hit_record: &HitRecord
) -> BSDFRecord

Return the result of a scattering function on an input ray

Loading content...