Module nib::sampler

The sampler interface

This module defines a generic sampling interface that can be extended to various sampling implementations.

Re-exports

pub use random::Random;

Modules

primitives
Sampling functions for various primitives

random
A basic random sampler

Enums

SamplerError
The possible errors that a Sampler can return

Traits

InPlace
The interface for an in-place sampler that can stochastically query samples

Sampler
The generic interface for a rendering sampler

Sequential
The interface for a sampler that needs to be queried sequentially

Type Definitions

SamplerResult
A convenient type alias for when a result can return a SamplerError