Trait nib::sampler::Sampler

pub trait Sampler<T>: Debug + Send + Sync where
    T: GenFloat, {
    fn sample_idx(&mut self, index: u32) -> Result<T, SamplerError<T>>;

    fn sampler_idx_dims(
        &mut self, 
        index: u32, 
        dimensions: u32
    ) -> Result<Vec<T>, SamplerError<T>>;

    fn next(&mut self, dimensions: u32) -> Result<Vec<T>, SamplerError<T>>;
}

The generic interface for a rendering sampler

This method automatically derives these methods for any sampler that implements the BaseSampler trait. Users should use this trait and not the BaseSampler trait as an interface.

Required methods

fn sample_idx(&mut self, index: u32) -> Result<T, SamplerError<T>>

Sample all of the dimensions for a particular index

fn sampler_idx_dims(
&mut self,
index: u32,
dimensions: u32
) -> Result<Vec<T>, SamplerError<T>>

Request a certain number of dimensions for a particular index

fn next(&mut self, dimensions: u32) -> Result<Vec<T>, SamplerError<T>>

Get the next dimensions samples.

If there are no more dimensions remaining for this particular index, then this will return an error, or an incomplete

Loading content...