Module nib::scene

A module providing definitions and methods for a scene. A scene contains most of the relevant information for an integrator, and it provides information such as lights, geometric objects, camera information, amongst other things. The scene primarily interacts with the deserializer and the integrator.

Structs

Scene
A struct representing the scene description as the user will input it

SerializedTextured
A serializable wrapper for the

Enums

SerializedAccelerationStruct
The different types of acceleration structures that can be used in the scene description

Constants

_IMPL_DESERIALIZE_FOR_Scene

_IMPL_DESERIALIZE_FOR_SerializedAccelerationStruct

_IMPL_DESERIALIZE_FOR_SerializedTextured

_IMPL_SERIALIZE_FOR_Scene

_IMPL_SERIALIZE_FOR_SerializedAccelerationStruct

_IMPL_SERIALIZE_FOR_SerializedTextured