Module nib::types

Defines generic numeric types for the integrator so that operations can be done with generic integers or floating point numbers.

Macros

aggregate_trait 🔒

Generate a trait that is the sum of other trait bounds

Constants

ETA

The floating point error threshold to use with the renderer

Traits

GenFloat

GenInteger

GenReal

Functions

approx_eq_vec

Compare two floating point vectors

eta

The floating point error threshold to use with the renderer, as a convenience function that will automatically convert it to whatever numeric type you want.

Type Aliases

Float

The particular floating point type that is going to be used in this program. If you want to switch the float type to another type, simply change the type here.

Integer

The particular integer type to use in this program.

PixelValue

A type representing the RGB value of a pixel in the rendering calculations. This is not the final color value that is output to the buffer.

Unsigned

The particular unsigned integer type to use in this program. To switch the int type to another type, just change the type here.