Trait nib::types::Preprocess

pub trait Preprocess {
    type Output;
    fn process(self) -> Self::Output;
}

Describes a struct that needs to be preprocessed before it is ready to be used at runtime.

Some structs need to be processed from the serialized input struct to be converted into something that can be used by the renderer. This is often used for objects in the renderer that have computations that can be precalculated and cached, or for structs that are serialized from user input and then converted into something else.

Every struct that is deserialized from user input must implement this trait. For your convenience, there is a default derive macro provided for items that don't require any procesing that simply returns the struct, it is equivalent to the following:

impl Preprocess for T {
    Output = T;

    fn process(self) -> Self::Output {
        self
    }
}

Usage is simple:

#[derive(Preprocess)]
struct Example {}

Associated Types

type Output

Loading content...
Required methods

fn process(self) -> Self::Output

Process a struct and return some new struct that is suitable for use in the renderer at runtime

Loading content...