nib::types
Type Alias PixelValue

pub type PixelValue<T> = Vector3<T>;

A type representing the RGB value of a pixel in the rendering calculations. This is not the final color value that is output to the buffer.

Aliased Type

struct PixelValue<T> {
    pub x: T,
    pub y: T,
    pub z: T,
}

Fields
x: T

The x component of the vector.

y: T

The y component of the vector.

z: T

The z component of the vector.