Trait nib::image_exporter::FramebufferExporterBase

pub trait FramebufferExporterBase {
    const MAX_COLOR: u32;

    fn export(
        &self, 
        buffer: &[Vector3<u32>], 
        path: &Path
    ) -> Result<(), ExporterError>;
}

The "base" trait for a FrameBufferExporter

Implementing this trait automatically implements the FrameBufferExporter trait, which provides some methods for free, such as converting the values from values between 0 and 1 to N-bit color values.

Associated Constants

const MAX_COLOR: u32

The maximum color value that a framebuffer

Loading content...
Required methods

fn export(
&self,
buffer: &[Vector3<u32>],
path: &Path
) -> Result<(), ExporterError>

Export a buffer of pixel values to a file

The buffer is a vector of RGB pixel values, and the path is the desired path to write the file. You can assume that the buffer will consist of integers between 0 and MAX_COLOR.

Loading content...