Trait nib::integrator::Integrator

pub trait Integrator<T: GenFloat>: Debug + Send + Sync {
    fn render(&self, params: RenderParams<T>) -> Vector3<T>;
}

A trait that defines an integrator. An integrator defines the operations that are responsible for taking input data for a given pixel, and calculating the output colors at each pixel.

Required methods

fn render(&self, params: RenderParams<T>) -> Vector3<T>

Calculate the color value for a particular pixel, given a reference to the scene.

Given certain input parameters this method calculates the color values at a particular point.

Loading content...